requirements.txt
pytest
fastapi
email-validator
python-dotenv
requests
pydantic
faker

psycopg2-binary
sqlalchemy
sqlmodel

fastapi-pagination